backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pymysql==1.1.0
marshmallow==3.20.1
email-validator==2.1.1
gunicorn==21.2.0
